+++ filamind_iot/models/__init__.py
@@ -2,4 +2,5 @@ from . import iot_device_type
 from . import iot_box
 from . import iot_device
 from . import iot_connection_log
+from . import iot_command_queue
 from . import iot_config_settings